Fixes a ton of size_t / int implicit conversion warnings
On my new version of clang

+++ src/ir.cpp
@@ -374,7 +374,7 @@ Stmt Print::make(std::string fmt, std::vector<Expr> params) {
 }
 
 // GetProperty
-Expr GetProperty::make(Expr tensor, TensorProperty property, int dim) {
+Expr GetProperty::make(Expr tensor, TensorProperty property, size_t dim) {
   if (property != TensorProperty::Values)
     iassert(dim >= 0) << "Must pass in which dimension you want property from.";
   GetProperty* gp = new GetProperty;
